website/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
sqlalchemy==2.0.23
stripe==7.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
requests==2.31.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
supabase==2.0.2
postgrest==0.13.2